+++ e2e/test/typescript-project-references/src/test/job.spec.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import {start} from '../src/core/job';
+import {start} from '../core/job';
 
 describe(start.name, () => {
   it('should format a correct message', () => {