+++ packages/core/src/initializer/custom-initializers/custom-initializer.ts
@@ -1,6 +1,11 @@
 import { PartialStrykerOptions } from '@stryker-mutator/api/core';
 
-export interface PresetConfiguration {
+export interface CustomInitializer {
+  readonly name: string;
+  createConfig(): Promise<CustomInitializerConfiguration>;
+}
+
+export interface CustomInitializerConfiguration {
   config: PartialStrykerOptions;
   guideUrl: string;
   dependencies: string[];