+++ packages/core/src/initializer/custom-initializers/react-initializer.ts
@@ -1,7 +1,6 @@
 import { StrykerOptions } from '@stryker-mutator/api/core';
 
-import { Preset } from './preset.js';
-import { PresetConfiguration } from './preset-configuration.js';
+import { CustomInitializer, CustomInitializerConfiguration } from './custom-initializer.js';
 
 const guideUrl = 'https://stryker-mutator.io/docs/stryker-js/guides/react';
 
@@ -9,7 +8,7 @@ const guideUrl = 'https://stryker-mutator.io/docs/stryker-js/guides/react';
  * More information can be found in the Stryker handbook:
  * https://stryker-mutator.io/docs/stryker-js/guides/react
  */
-export class ReactPreset implements Preset {
+export class ReactInitializer implements CustomInitializer {
   public readonly name = 'create-react-app';
   private readonly dependencies = ['@stryker-mutator/jest-runner'];
 
@@ -22,7 +21,7 @@ export class ReactPreset implements Preset {
     },
   };
 
-  public async createConfig(): Promise<PresetConfiguration> {
+  public async createConfig(): Promise<CustomInitializerConfiguration> {
     return { config: this.config, guideUrl, dependencies: this.dependencies };
   }
 }